Change behavior on default values
Upon default value retrieval, the old values are kept.
Datarate delay calculation is now done in 64bit instead of double values.
Cell ids can be retrieved from the database

+++ src/queues/implementation/datarate/TbusDatarateQueue.h
@@ -30,7 +30,7 @@ class TbusDatarateQueue : public TbusBaseQueue<TbusQueueDatarateValue> {
 		TbusDatarateQueue(TbusQueueSelection selection);
 
 	protected:
-		double bitsSent; ///< Bits already sent for head of queue (Needed for multiple data- and droprate value per packet)
+		int64_t bitsSent; ///< Bits already sent for head of queue (Needed for multiple data- and droprate value per packet)
 
 		virtual void calculateEarliestDeliveries();
 		virtual void calculateEarliestDeliveryForPacket(cPacket* packet);
